Fix: removing unused states
